[TASK] Added semi secure header

+++ models/db/cumulated_orders.go
@@ -4,7 +4,7 @@ import "github.com/jinzhu/gorm"
 
 type CumulatedOrders struct {
 	gorm.Model
-	Team            Team  `gorm:"foreignkey:TeamID" json:"-"`
-	TeamID          uint  `json:"-" `
-	CumulatedPoints int64 `json:"cumulated_points" `
+	Team            Team    `gorm:"foreignkey:TeamID" json:"-"`
+	TeamID          uint    `json:"-" `
+	CumulatedPoints float64 `json:"cumulated_points" `
 }
