+++ models/db/order.go
@@ -4,7 +4,7 @@ import "github.com/jinzhu/gorm"
 
 type Order struct {
 	gorm.Model
-	Team   Team `gorm:"foreignkey:TeamID"`
-	TeamID uint `json:"teamid" `
-	Points int  `json:"points" `
+	Team   Team    `gorm:"foreignkey:TeamID"`
+	TeamID uint    `json:"teamid" `
+	Points float64 `json:"points" `
 }
